refactor: split extensions into ast_ext for better DX
