+++ crates/rolldown_ecmascript_utils/src/extensions/allocator_ext.rs
