+++ crates/rolldown_ecmascript_utils/src/extensions/ast_ext/binding_pattern_ext.rs
@@ -1,25 +1,11 @@
-use oxc::allocator::{Allocator, Box, IntoIn};
-use oxc::ast::ast;
-use oxc::semantic::SymbolId;
-use oxc::span::SPAN;
+use oxc::{
+  allocator::{Allocator, Box, IntoIn},
+  ast::ast,
+  span::SPAN,
+};
 use smallvec::SmallVec;
 
-use crate::allocator_helpers::take_in::TakeIn;
-use crate::AstSnippet;
-
-pub mod allocator_ext;
-pub mod span_ext;
-
-pub trait BindingIdentifierExt {
-  fn expect_symbol_id(&self) -> SymbolId;
-}
-
-impl BindingIdentifierExt for ast::BindingIdentifier<'_> {
-  #[inline]
-  fn expect_symbol_id(&self) -> SymbolId {
-    self.symbol_id.get().unwrap_or_else(|| panic!("fail get symbol id from {self:?}"))
-  }
-}
+use crate::{AstSnippet, TakeIn};
 
 pub trait BindingPatternExt<'ast> {
   fn binding_identifiers(&self) -> smallvec::SmallVec<[&Box<ast::BindingIdentifier<'ast>>; 1]>;
@@ -173,162 +159,5 @@ impl<'ast> BindingPatternExt<'ast> for ast::BindingPattern<'ast> {
       }
     };
     left
-  }
-}
-
-pub trait StatementExt<'me, 'ast> {
-  fn is_import_declaration(&self) -> bool;
-  fn as_import_declaration(&'me self) -> Option<&'me ast::ImportDeclaration<'ast>>;
-  fn as_export_default_declaration_mut(
-    &'me mut self,
-  ) -> Option<&'me mut ast::ExportDefaultDeclaration<'ast>>;
-  fn as_export_all_declaration(&self) -> Option<&ast::ExportAllDeclaration<'ast>>;
-  fn as_export_named_declaration(&self) -> Option<&ast::ExportNamedDeclaration<'ast>>;
-  fn as_export_named_declaration_mut(&mut self) -> Option<&mut ast::ExportNamedDeclaration<'ast>>;
-
-  fn is_function_declaration(&self) -> bool;
-  fn as_function_declaration(&self) -> Option<&ast::Function<'ast>>;
-
-  fn is_module_declaration_with_source(&self) -> bool;
-}
-
-impl<'ast> StatementExt<'_, 'ast> for ast::Statement<'ast> {
-  fn is_import_declaration(&self) -> bool {
-    matches!(self, ast::Statement::ImportDeclaration(_))
-  }
-
-  fn as_import_declaration(&self) -> Option<&ast::ImportDeclaration<'ast>> {
-    if let ast::Statement::ImportDeclaration(import_decl) = self {
-      return Some(&**import_decl);
-    }
-    None
-  }
-
-  fn as_export_default_declaration_mut(
-    &mut self,
-  ) -> Option<&mut ast::ExportDefaultDeclaration<'ast>> {
-    if let ast::Statement::ExportDefaultDeclaration(export_default_decl) = self {
-      return Some(&mut **export_default_decl);
-    }
-    None
-  }
-
-  fn as_export_all_declaration(&self) -> Option<&ast::ExportAllDeclaration<'ast>> {
-    if let ast::Statement::ExportAllDeclaration(export_all_decl) = self {
-      return Some(&**export_all_decl);
-    }
-    None
-  }
-
-  fn as_export_named_declaration(&self) -> Option<&ast::ExportNamedDeclaration<'ast>> {
-    if let ast::Statement::ExportNamedDeclaration(export_named_decl) = self {
-      return Some(&**export_named_decl);
-    }
-    None
-  }
-
-  fn as_export_named_declaration_mut(&mut self) -> Option<&mut ast::ExportNamedDeclaration<'ast>> {
-    if let ast::Statement::ExportNamedDeclaration(export_named_decl) = self {
-      return Some(&mut **export_named_decl);
-    }
-    None
-  }
-
-  fn as_function_declaration(&self) -> Option<&ast::Function<'ast>> {
-    if let ast::Statement::FunctionDeclaration(func_decl) = self {
-      Some(func_decl)
-    } else {
-      None
-    }
-  }
-
-  fn is_function_declaration(&self) -> bool {
-    self.as_function_declaration().is_some()
-  }
-
-  /// Check if the statement is `[import|export] ... from ...` or `export ... from ...`
-  fn is_module_declaration_with_source(&self) -> bool {
-    matches!(self.as_module_declaration(), Some(decl) if decl.source().is_some())
-  }
-}
-
-pub trait ExpressionExt<'ast> {
-  fn as_call_expression(&self) -> Option<&ast::CallExpression<'ast>>;
-  fn as_call_expression_mut(&mut self) -> Option<&mut ast::CallExpression<'ast>>;
-
-  fn as_identifier(&self) -> Option<&ast::IdentifierReference<'ast>>;
-  fn as_identifier_mut(&mut self) -> Option<&mut ast::IdentifierReference<'ast>>;
-  fn as_unary_expression(&self) -> Option<&ast::UnaryExpression<'ast>>;
-  fn as_string_literal(&self) -> Option<&ast::StringLiteral<'ast>>;
-  fn as_binary_expression(&self) -> Option<&ast::BinaryExpression<'ast>>;
-
-  fn is_import_meta(&self) -> bool;
-  fn is_import_meta_url(&self) -> bool;
-}
-
-impl<'ast> ExpressionExt<'ast> for ast::Expression<'ast> {
-  fn as_call_expression(&self) -> Option<&ast::CallExpression<'ast>> {
-    if let ast::Expression::CallExpression(call_expr) = self {
-      Some(call_expr)
-    } else {
-      None
-    }
-  }
-
-  fn as_call_expression_mut(&mut self) -> Option<&mut ast::CallExpression<'ast>> {
-    if let ast::Expression::CallExpression(call_expr) = self {
-      Some(call_expr)
-    } else {
-      None
-    }
-  }
-
-  fn as_identifier(&self) -> Option<&ast::IdentifierReference<'ast>> {
-    if let ast::Expression::Identifier(ident) = self {
-      Some(ident)
-    } else {
-      None
-    }
-  }
-
-  fn as_identifier_mut(&mut self) -> Option<&mut ast::IdentifierReference<'ast>> {
-    if let ast::Expression::Identifier(ident) = self {
-      Some(ident)
-    } else {
-      None
-    }
-  }
-
-  fn as_unary_expression(&self) -> Option<&ast::UnaryExpression<'ast>> {
-    let ast::Expression::UnaryExpression(expr) = self else {
-      return None;
-    };
-    Some(expr)
-  }
-
-  fn as_string_literal(&self) -> Option<&ast::StringLiteral<'ast>> {
-    let ast::Expression::StringLiteral(expr) = self else {
-      return None;
-    };
-    Some(expr)
-  }
-
-  fn as_binary_expression(&self) -> Option<&ast::BinaryExpression<'ast>> {
-    let ast::Expression::BinaryExpression(expr) = self else {
-      return None;
-    };
-    Some(expr)
-  }
-
-  /// // Check if the expression is `import.meta`
-  fn is_import_meta(&self) -> bool {
-    matches!(self, ast::Expression::MetaProperty(meta_prop)
-    if meta_prop.meta.name == "import" && meta_prop.property.name == "meta")
-  }
-
-  /// Check if the expression is `import.meta.url`
-  fn is_import_meta_url(&self) -> bool {
-    matches!(self, ast::Expression::StaticMemberExpression(member_expr)
-    if member_expr.object.is_import_meta() && member_expr.property.name == "url")
   }
 }
