+++ crates/rolldown_ecmascript_utils/src/extensions/ast_ext/mod.rs
@@ -1,0 +1,4 @@
+pub mod binding_identifier_ext;
+pub mod binding_pattern_ext;
+pub mod expression_ext;
+pub mod statement_ext;
