+++ crates/rolldown_ecmascript_utils/src/extensions/mod.rs
@@ -1,0 +1,3 @@
+pub mod allocator_ext;
+pub mod ast_ext;
+pub mod span_ext;
