+++ crates/rolldown_ecmascript_utils/src/extensions/span_ext.rs
