+++ crates/rolldown_ecmascript_utils/src/lib.rs
@@ -1,12 +1,16 @@
 mod allocator_helpers;
 mod ast_snippet;
-mod ext;
+mod extensions;
 
 pub use crate::{
   allocator_helpers::take_in::TakeIn,
   ast_snippet::AstSnippet,
-  ext::{
-    allocator_ext::AllocatorExt, span_ext::SpanExt, BindingIdentifierExt, BindingPatternExt,
-    ExpressionExt, StatementExt,
+  extensions::{
+    allocator_ext::AllocatorExt,
+    ast_ext::{
+      binding_identifier_ext::BindingIdentifierExt, binding_pattern_ext::BindingPatternExt,
+      expression_ext::ExpressionExt, statement_ext::StatementExt,
+    },
+    span_ext::SpanExt,
   },
 };
